refactoring to use the suggested approach for location & moving code into its own module
